+++ spec/system/admin/adding_to_an_order_spec.rb
@@ -212,86 +212,6 @@ RSpec.describe "Adding to an existing order" do
         expect(page).to have_link("Make a Reservation")
       end
     end
-
-    describe "from another facility", :js, feature_setting: { cross_core_projects: true } do
-      let(:facility2) { create(:setup_facility) }
-      let(:product) { create(:setup_item, :with_facility_account) }
-      let!(:instrument) { create(:setup_instrument, facility: facility2, cross_core_ordering_available: true) }
-      let(:user) { create(:user, :facility_administrator, facility:) }
-      let!(:facility2_account) { create(:account, :with_account_owner, type: "CreditCardAccount", owner: order.user, description: "Other Account", facility: facility2) }
-
-      describe "with one reservation" do
-        before do
-          visit facility_order_path(facility, order)
-          select_from_chosen facility2.name, from: "add_to_order_form[facility_id]"
-          select_from_chosen instrument.name, from: "add_to_order_form[product_id]"
-          fill_in "add_to_order_form[quantity]", with: "1"
-          select_from_chosen facility2_account.to_s, from: "Payment Source", scroll_to: :center
-          click_button "Add to Cross-Core Order"
-        end
-
-        it "requires a reservation to be set up before adding to the order" do
-          expect(page).to have_content("Your order includes one or more incomplete items.")
-
-          within("#merge-order-missing-order-info-reminder") do
-            click_link "Make a Reservation"
-          end
-          select facility2_account.to_s, from: "Payment Source"
-          click_button "Create"
-
-          expect(order.reload.order_details.count).to be(1)
-          project = order.cross_core_project
-          expect(project).to be_present
-          expect(project.orders.last.order_details.last.product).to eq(instrument)
-          expect(project.orders.last.account_id).to eq(facility2_account.id)
-        end
-
-        it "brings you back to the facility order path on 'Cancel'" do
-
-          within("#merge-order-missing-order-info-reminder") do
-            click_link "Make a Reservation"
-          end
-          click_link "Cancel"
-
-          expect(current_path).to eq(facility_order_path(facility, order))
-          expect(page).to have_link("Make a Reservation")
-        end
-      end
-
-      describe "with a product from before" do
-        let(:facility2) { create(:setup_facility) }
-        let!(:product2) { create(:setup_item, facility: facility2, cross_core_ordering_available: true) }
-
-        before do
-          visit facility_order_path(facility, order)
-          select_from_chosen facility2.name, from: "add_to_order_form[facility_id]"
-          select_from_chosen product2.name, from: "add_to_order_form[product_id]"
-          fill_in "add_to_order_form[quantity]", with: "1"
-          click_button "Add to Cross-Core Order"
-        end
-
-        it "sets the merge_with_order_id until the reservation is created" do
-          project = order.reload.cross_core_project
-          second_facility_order = project.orders.last
-
-          select_from_chosen facility2.name, from: "add_to_order_form[facility_id]", scroll_to: :center
-          select_from_chosen instrument.name, from: "add_to_order_form[product_id]"
-          fill_in "add_to_order_form[quantity]", with: "1"
-          click_button "Add to Cross-Core Order"
-
-          # This is the second order for this facility so it has a merge_order set
-          expect(project.reload.orders.last.merge_with_order_id).to eq(second_facility_order.id)
-
-          within("#merge-order-missing-order-info-reminder") do
-            click_link "Make a Reservation"
-          end
-          click_button "Create"
-
-          # After the reservation is created, the merge_order is cleared
-          expect(project.reload.orders.last.merge_with_order_id).to eq(nil)
-        end
-      end
-    end
   end
 
   describe "adding a timed service" do
@@ -310,103 +230,4 @@ RSpec.describe "Adding to an existing order" do
       expect(order.order_details.last.quantity).to eq(31)
     end
   end
-
-  describe "adding a product from another facility", :js, feature_setting: { cross_core_projects: true } do
-    let(:facility2) { create(:setup_facility) }
-    let(:product) { create(:setup_item, :with_facility_account, cross_core_ordering_available: false) }
-    let!(:product2) { create(:setup_item, :with_facility_account, facility: facility2, cross_core_ordering_available: false) }
-    let!(:cross_core_product_facility) { create(:setup_item, :with_facility_account, facility:, cross_core_ordering_available: true) }
-    let!(:cross_core_product_facility2) { create(:setup_item, :with_facility_account, facility: facility2, cross_core_ordering_available: true) }
-
-    before do
-      visit facility_order_path(facility, order)
-    end
-
-    context "with staff role" do
-      let!(:facility2_account) { create(:account, :with_account_owner, type: "CreditCardAccount", owner: order.user, description: "Other Account", facility: facility2) }
-
-      it "changes the button text" do
-        expect(page).to have_button("Add To Order")
-        select_from_chosen facility2.name, from: "add_to_order_form[facility_id]"
-        select_from_chosen cross_core_product_facility2.name, from: "add_to_order_form[product_id]"
-        expect(page).to have_button("Add to Cross-Core Order")
-      end
-
-      it "creates a new order for the selected facility" do
-        expect(page).not_to have_content("Cross Core Project ID")
-        expect(page.has_selector?("option", text: cross_core_product_facility.name, visible: false)).to be(true)
-        expect(page.has_selector?("option", text: product.name, visible: false)).to be(true)
-        expect(page.has_selector?("option", text: facility2_account.to_s, visible: false)).to be(false)
-
-        select_from_chosen facility2.name, from: "add_to_order_form[facility_id]", scroll_to: :center
-        select_from_chosen cross_core_product_facility2.name, from: "add_to_order_form[product_id]"
-        select_from_chosen facility2_account.to_s, from: "Payment Source", scroll_to: :center
-
-        expect(page.has_selector?("option", text: product2.name, visible: false)).to be(false)
-
-        click_button "Add to Cross-Core Order"
-
-        expect(page).to have_content("#{cross_core_product_facility2.name} was successfully added to this order.")
-        expect(page).to have_content(facility2.to_s), count: 2
-        expect(page).to have_content(user.full_name), count: 2
-
-        order.reload
-
-        project = order.cross_core_project
-        expect(project).to be_present
-
-        expect(page).to have_content("Cross-Core Project ID")
-        expect(page).to have_content(project.id)
-
-        project_total = project.orders.sum(&:total)
-        expect(page).to have_content("Cross-Core Project Total")
-        expect(page).to have_content(project_total)
-      end
-    end
-
-    context "with admin role" do
-      let(:user) { create(:user, :facility_administrator, facility:) }
-      let!(:facility2_account) { create(:account, :with_account_owner, type: "CreditCardAccount", owner: order.user, description: "Other Account", facility: facility2) }
-
-      it "changes the button text" do
-        expect(page).to have_button("Add To Order")
-        select_from_chosen facility2.name, from: "add_to_order_form[facility_id]"
-        select_from_chosen cross_core_product_facility2.name, from: "add_to_order_form[product_id]"
-        expect(page).to have_button("Add to Cross-Core Order")
-      end
-
-      it "creates a new order for the selected facility" do
-        expect(page).not_to have_content("Cross Core Project ID")
-        expect(page.has_selector?("option", text: cross_core_product_facility.name, visible: false)).to be(true)
-        expect(page.has_selector?("option", text: product.name, visible: false)).to be(true)
-        expect(page.has_selector?("option", text: facility2_account.to_s, visible: false)).to be(false)
-
-        select_from_chosen facility2.name, from: "add_to_order_form[facility_id]", scroll_to: :center
-        select_from_chosen cross_core_product_facility2.name, from: "add_to_order_form[product_id]"
-        select_from_chosen facility2_account.to_s, from: "Payment Source", scroll_to: :center
-
-        expect(page.has_selector?("option", text: product2.name, visible: false)).to be(false)
-
-        click_button "Add to Cross-Core Order"
-
-        expect(page).to have_content("#{cross_core_product_facility2.name} was successfully added to this order.")
-        expect(page).to have_content(facility2.to_s), count: 2
-        expect(page).to have_content(user.full_name), count: 2
-
-        order.reload
-
-        project = order.cross_core_project
-        expect(project).to be_present
-
-        expect(page).to have_content("Cross-Core Project ID")
-        expect(page).to have_content(project.id)
-
-        project_total = project.orders.sum(&:total)
-        expect(page).to have_content("Cross-Core Project Total")
-        expect(page).to have_content(project_total)
-      end
-    end
-
-  end
-
 end
